adding IsHelm helper function
