+++ pkg/safeguards/tests/testmanifests/nochart-invalid/values.yaml
@@ -1,0 +1,12 @@
+replicaCount: 1
+image:
+  repository: nginx
+  tag: stable
+service:
+  type: ClusterIP
+  port: 80
+ingress:
+  enabled: true
+  hostname: example.com
+releaseName: test-release
+releaseNamespace: test-namespace
